Dynamically display ingredients based on selected recipe.
This PR is worth re-learning though because it has a lot of tricks
on sending data from widget configuration to widget class to service
to provider.

Nevertheless, this works ONLY when adding the first widget.
Weirdly when adding the second widget, this second widget still
pulls data from the first widget's selected recipe ID.

It's a BUG and I need to figure it out.